+++ shared/lib/server-consts.js
@@ -29,6 +29,8 @@ var ServerConsts;
         LOG_LEVEL_SUFFIX: 'loglevel=',
         TELEPORT_SUFFIX: 'teleport=',
         RENDER_VAR_SUFFIX: 'render-var=',
+        WIDTH_SUFFIX: 'width=',
+        HEIGHT_SUFFIX: 'height=',
         // client messages
         MSG_READY: 'ready',
         MSG_RENDER: 'render',