+++ shared/lib/server-consts.ts
@@ -36,6 +36,8 @@ export namespace ServerConsts {
     LOG_LEVEL_SUFFIX: 'loglevel=',
     TELEPORT_SUFFIX: 'teleport=',
     RENDER_VAR_SUFFIX: 'render-var=',
+    WIDTH_SUFFIX: 'width=',
+    HEIGHT_SUFFIX: 'height=',
 
     // client messages
     MSG_READY: 'ready',